# Conflicts:
#	src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/prepost/flow_modelinit.F90

+++ src/engines_gpl/rr/packages/rr_kernel_f/src/BinFileModule.f90
@@ -101,7 +101,7 @@ contains
     Integer     teller, iOW, iCall, in
     Integer     iNode, iKind, iNr
     Character(CharIdLength) idNode, FileName
-    Integer iDebug, iOut1, infile1, infile2, infile3, infile4, infile5, infile6
+    Integer(4) iDebug, iOut1, infile1, infile2, infile3, infile4, infile5, infile6
     Integer SetMaxTabNr, SetMaxDataNr
     Character(Len=FilCharIdLength)            :: boundCondFile
